Hand written concepts replaced with Casey Carter's repository

+++ example/example.cpp
@@ -30,7 +30,7 @@ void foo(V v, T t)
 {
   const Length distance = v * t;
   std::cout << "A car driving " << v.count() << " km/h in a time of " << t.count() << " minutes will pass "
-            << quantity_cast<length<meter, int>>(distance).count() << " meters.\n";
+            << quantity_cast<length<meter, double>>(distance).count() << " meters.\n";
 }
 
 void foo()
